feat: CLIN-4328 - etl_run
