+++ dags/lib/utils_etl.py
@@ -1,7 +1,9 @@
 import json
 from enum import Enum
-from typing import List, Optional
+from typing import Dict, List, Optional
 
+from airflow.decorators import task
+from airflow.models import DagRun
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
 from lib import config
 from lib.config import clin_import_bucket, config_file
@@ -131,3 +133,36 @@ def build_etl_job_arguments(
     if chromosome:
         arguments = arguments + ['--chromosome', f'chr{chromosome}']
     return arguments
+
+@task(task_id='get_ingest_dag_configs_by_batch_id')
+def get_ingest_dag_configs_by_batch_id(batch_id: str, ti=None) -> dict:
+    dag_run: DagRun = ti.dag_run
+    return {
+        'batch_id': batch_id,
+        'sequencing_ids': None,
+        'color': dag_run.conf['color'],
+        'import': dag_run.conf['import'],
+        'spark_jar': dag_run.conf['spark_jar']
+    }
+    
+@task(task_id='get_ingest_dag_configs_by_sequencing_ids')
+def get_ingest_dag_configs_by_sequencing_ids(all_batch_types: Dict[str, str], sequencing_ids: List[str], analysisType: str, ti=None) -> dict:
+    dag_run: DagRun = ti.dag_run
+
+    # try regroup sequencing ids and generate a config of etl_ingest for each analysis type
+
+    sequencing_ids_compatible_with_type = []
+    for identifier, type in all_batch_types.items():
+        if analysisType == type and identifier in sequencing_ids:
+            sequencing_ids_compatible_with_type.append(identifier)
+
+    if len(sequencing_ids_compatible_with_type) == 0:
+        return None # No sequencing ids found for that analysis type
+
+    return {
+        'batch_id': None,
+        'sequencing_ids': sequencing_ids_compatible_with_type,
+        'color': dag_run.conf['color'],
+        'import': dag_run.conf['import'],
+        'spark_jar': dag_run.conf['spark_jar']
+    }
